% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.segmentation.R
\name{aggregate.segments}
\alias{aggregate.segments}
\title{Aggregate positive and negative segments}
\usage{
\method{aggregate}{segments}(
  cna.matrix,
  map.loc.agr,
  amp.level,
  del.level,
  params.p,
  params.n,
  fdr,
  samps.use = vector("numeric")
)
}
\arguments{
\item{cna.matrix}{Copy number matrix}

\item{map.loc.agr}{Aggregated map locations}

\item{amp.level}{Amplification threshold}

\item{del.level}{Deletion threshold}

\item{params.p}{Positive parameters}

\item{params.n}{Negative parameters}

\item{fdr}{False discovery rate}

\item{samps.use}{Samples to use}
}
\value{
List containing positive and negative segments
}
\description{
Aggregate positive and negative segments
}
