% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{Gene locations}
\format{\code{\link[data.table]{data.table}} with 5 columns
\describe{
\item{ID}{Gene ID}
\item{Name}{Gene symbol}
\item{Chromosome}{Chromosome name}
\item{Start}{Gene start}
\item{End}{Gene end}
}}
\usage{
genes
}
\description{
Gene locations used by RUBIC example to call focal events.
}
\examples{
  genes
}
\keyword{datasets}

