% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers}
\alias{markers}
\title{Markers locations}
\format{\code{\link[data.table]{data.table}} with 3 columns
\describe{
\item{Name}{Probe name}
\item{Chromosome}{Chromosome name}
\item{Position}{Probe position on the chromosome}
}}
\usage{
markers
}
\description{
The markers locations used by RUBIC example that indicates the exact
locations of measurement probes (markers) for the given platform.
For sequencing data, copy number values are often estimated with fixed
bin sizes (prior to segmentations). In this case each marker should be
associated with a bin and the center genomic position of the bin.
}
\examples{
  markers
}
\keyword{datasets}

