% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.events.R
\name{read.genes.info.biomart}
\alias{read.genes.info.biomart}
\title{Import Biomart annotations from the web service.}
\usage{
read.genes.info.biomart(filter.chr = NULL)
}
\arguments{
\item{filter.chr}{A character vector of wanted chromosomes.}
}
\value{
The \code{data.table} containing the genome annotations.
}
\description{
Import Biomart annotations from the web service.
}

