% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.events.R
\name{read.genes.info.tsv}
\alias{read.genes.info.tsv}
\title{Read Biomart annotation file.}
\usage{
read.genes.info.tsv(build.file, filter.chr = NULL)
}
\arguments{
\item{build.file}{The TSV file exported from Biomart.}

\item{filter.chr}{A character vector of wanted chromosomes.}
}
\value{
The \code{data.table} containing the genome annotations.
}
\description{
Read Biomart annotation file.
}
\section{Warning}{

This function is expecting at least the following headers: 'Ensembl Gene ID',
'Gene Start (bp)', 'Gene End (bp)', 'Chromosome Name', 'Associated Gene Name'.
}

