% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{read.seg.file}
\alias{read.seg.file}
\title{Read the segmented CNA file in TSV format.}
\usage{
read.seg.file(seg.file, sample = 1, chromosome = 2, start = 3, end = 4,
  log.ratio = 6, header = T)
}
\arguments{
\item{seg.file}{The name of the file containig the segmented CNA.}

\item{sample}{The index of the sample name column in the \code{seg.file}.}

\item{chromosome}{The index of the chromosome column in the \code{seg.file}.}

\item{start}{The index of the segment start column in the \code{seg.file}.}

\item{end}{The index of the segment end column in the \code{seg.file}.}

\item{log.ratio}{The index of the segment log ratio column in the \code{seg.file}.}

\item{header}{A boolean value that specifies whether the file contain a header line.}
}
\value{
A \code{data.table} containing the segmented CNA.
}
\description{
Read the segmented CNA file in TSV format.
}

