% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubic.R
\docType{class}
\name{RUBIC-class}
\alias{RUBIC}
\alias{RUBIC-class}
\title{RUBIC Class}
\description{
A ReferenceClass that contains all the data needed and generated by the RUBIC algorithm.
}
\details{
It is possibile and recommended to instanciate a new RUBIC object by using the
         convenience function \code{\link{rubic}}.
}
\section{Fields}{

\describe{
\item{\code{fdr}}{Event based false descovery rate.}

\item{\code{genes}}{A data.table containg the gene locations.}

\item{\code{samples}}{A character vector containing sample IDs.}

\item{\code{markers}}{A data.table containing the markers informations.}

\item{\code{map.loc}}{A data table containing the markers mapped on the segments.}

\item{\code{min.mean}}{The minimum mean copy number allowed.}

\item{\code{max.mean}}{The maximum mean copy number allowed.}

\item{\code{min.probes}}{The minimum number of probes to be considered in the analysis.}

\item{\code{min.seg.markers}}{The number of probes allowed in each segment.}

\item{\code{amp.level}}{The threshold used for calling amplifications.}

\item{\code{del.level}}{The threshold used for calling deletions.}

\item{\code{focal.threshold}}{The maximum length of a recurrent region to be called focal.}
}}
\section{Methods}{

\describe{
\item{\code{call.events()}}{Call recurrent events.}

\item{\code{call.focal.events(genes = NULL)}}{Call focal events.}

\item{\code{estimate.parameters(quiet = T, test.env = NULL)}}{Estimate the parameters necessary for segmentation and event calling.}

\item{\code{initialize(fdr = NULL, seg.cna = NULL, markers = NULL, samples = NULL,
  genes = NULL, amp.level = 0.1, del.level = -0.1, min.seg.markers = 1,
  min.mean = NA_real_, max.mean = NA_real_, min.probes = 260000,
  focal.threshold = 1e+07, seg.cna.header = T, markers.header = T,
  samples.header = F, col.sample = 1, col.chromosome = 2, col.start = 3,
  col.end = 4, col.log.ratio = 6, ...)}}{Create a new RUBIC object. It is recommended to use the \link{rubic} function instead
of calling this method directly.
See \link{rubic} for parameter descriptions.}

\item{\code{save(file)}}{Save the current state of the RUBIC object to file.}

\item{\code{save.focal.gains(file)}}{Save focal gains to a file in TSV format.}

\item{\code{save.focal.losses(file)}}{Save focal losses to a file in TSV format.}

\item{\code{save.plots(dir, genes = NULL, steps = T, width = 11, height = 5)}}{Save gains and losses plots for each chromosome.}

\item{\code{segment()}}{Generate positive and negative segments.}
}}
\seealso{
\code{\link{rubic}}
}

