% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples}
\alias{samples}
\title{Samples names}
\format{\code{\link[data.table]{data.table}} with 1 column
\describe{
\item{V1}{Sample name}
}}
\usage{
samples
}
\description{
Samples names used to filter samples in the \code{\link{seg.cna}} data for
the RUBIC example.
}
\examples{
  samples
}
\keyword{datasets}

