% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seg.cna}
\alias{seg.cna}
\title{Segmented CNA}
\format{\code{\link[data.table]{data.table}} with 5 columns
\describe{
\item{Sample}{Sample anme}
\item{Chromosome}{Chromosome name}
\item{Start}{Segment start}
\item{End}{Segment end}
\item{LogRatio}{Log ratios for this segment}
}}
\usage{
seg.cna
}
\description{
Segmented CNA simulated data used by RUBIC example.
}
\examples{
  seg.cna
}
\keyword{datasets}

