% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.events.R
\name{sort.regions.on.genome}
\alias{sort.regions.on.genome}
\title{Sort focal events by genomic position}
\usage{
\method{sort}{regions.on.genome}(focal.events)
}
\arguments{
\item{focal.events}{List of focal events to be sorted}
}
\value{
Sorted list of focal events
}
\description{
Sort focal events by genomic position
}
